split the OSC library and removed wrapper, new examples
yes htis probably should be more than one commit.
