--- processing/examples/oscSerialBasics/code/OscSerial.jar
@@ -1,1 +1,0 @@
-../../../library/OscSerial.jar
